refactor: Remove `execa` dependency
When we merged `@guardian/cdk-cli` into this repo, we created two ways of creating a child process:
  - Manually using `child_process`
  - Using the `execa` library

This change DRYs the repo out a little by removing the `execa` dependency.

It also co-locates the git commands.
