+++ package.json
@@ -76,7 +76,6 @@
     "chalk": "^4.1.2",
     "cli-ux": "^5.6.6",
     "codemaker": "^1.54.0",
-    "execa": "^5.1.1",
     "git-url-parse": "^11.6.0",
     "inquirer": "^8.2.0",
     "lodash.camelcase": "^4.3.0",
